Проведена работа с замыканиями, были внесены изменения в MovieQuizViewController в метод показа результатов квиза и диспетчер задач добавлены слабые ссылки. Разделили и вынесли ответственность, связали логику приложения.
