+++ MovieQuiz/Services/QuestionFactoryProtocol.swift
@@ -1,0 +1,11 @@
+//
+//  QuestionFactoryProtocol.swift
+//  MovieQuiz
+//
+//  Created by Чингиз Джабаев on 18.07.2023.
+//
+
+import Foundation
+protocol QuestionFactoryProtocol {
+    func requestNextQuestion() -> QuizQuestion?
+}
